pep440: add tests for human visible error messages
Prior to the refactoring in previous commits, tests were generally
written against the rendered error message instead of the structured
message. This was in part because many of the error types themselves
were just a `String`, but also to explicitly test what an end user
actually sees. That's valuable.

While we keep most of the tests as rewritten to target the new
structured error representation, we add some new tests that captures the
value of testing the messages than humans will actually see.