pep440: shrink the size of Version
This puts a `Version` inside of an `Arc` internally. This "works" pretty
transparently because `Version` does not expose any methods for direct
in-place mutation.

While this doesn't *obviously* speed anything up directly, it does
result in a fairly sizeable improvement to the `boto3` requirements
list:

    $ time puffin-cmp pip-compile --no-build --cache-dir ~/astral/tmp/cache/ -o /dev/null ./scripts/requirements/boto3.in
    Resolved 31 packages in 10.84s

    real    10.858
    user    10.272
    sys     0.444
    maxmem  2402 MB
    faults  0

    $ time puffin-cmparc pip-compile --no-build --cache-dir ~/astral/tmp/cache/ -o /dev/null ./scripts/requirements/boto3.in
    Resolved 31 packages in 9.20s

    real    9.218
    user    8.919
    sys     0.165
    maxmem  463 MB
    faults  0

My hypothesis for why this leads to an improvement is that this shrinks
the size of a `Version` from 104 bytes(!) to 8 bytes. Since it looks
like pubgrub builds in-memory collections of `Version` values, this in
turn substantially shrinks heap usage of those collections and also
likely makes things more cache friendly.

It's plausible that an even better approach here would be to only wrap a
`VersionFull` in an `Arc`. But to make this most effective, we'd want to
truly represent a `VersionSmall` as just a single `u64`. Perhaps even a
`NonZeroU64` to create a niche. But to do that, I think we need to:

Represent the number of release segments in `VersionSmall::repr`. This
should be easy enough to do, I think, by commandeering the two least
significant bits of the release component (currently using the 5 most
significant bytes of the u64 representation). This shouldn't impact
comparisons since the lengths are at the least significant part, and
thus, will be equivalent if a comparison needs to consider them. If we
borrow the two bits from the first release segment (which is a u16),
then it should still be able to represent calver since 2^14=16384.

Then `Version` could be defined like this:

    struct Version(VersionInner);

    enum VersionInner {
        Small(u64),
        Full(Arc<VersionFull>),
    }

Although this does raise the size of `Version` from the 8 bytes in this
PR to 16. One might be able to reduce it back to 8 with a niche
optimization, but I can't think of any obvious way of doing that in safe
code.