Implement `--find-links` as flat indexes (pip-compile)
Add directory `--find-links` support for local paths to pip-compile.

It seems that pip joins all sources and then picks the best package. We explicitly give find links packages precedence if the same exists on an index and locally by prefilling the `VersionMap`, otherwise they are added as another index and the existing rules of precedence apply.

Internally, the feature is called _flat index_, which is more meaningful than _find links_: We're not looking for links, we're picking up local directories, and (TBD) support another index format that's just a flat list of files instead of a nested index.

Previously, a `RegistryBuiltDist`/`RegistrySourceDist` would be formatted as `<name>==<version>` while path and url dists would be formatted as `<name> @ <url>`. Now there are flat index path and url dists that behave as additions to the index and should be formatted by version instead of by url. Instead of adding more `Dist` types, i chose to track their designation as `is_registry`. An alternative would be adding an `is_registry` field to path and url dists.

`PrioritizedDistribution` and `FlatIndex` have been moved to locations where we can use them in the upstack PR.

I added a `scripts/wheels` directory with stripped down wheels to use for testing.

We're lacking tests for correct tag priority precedence with flat indexes, i only confirmed this manually since it is not covered in the pip-compile or pip-sync output.

Part 1/3 (directory find links in pip-compile, find links in pip-sync, remote find links).