pep440: (PreReleaseKind, u64) -> PreRelease
This turns our tuple of pre-release kind and number into a named type.

The reason for doing this is a little subtle, but with the impending
attempt to use zero-copy deserialization, we need all of our types that
live in a cache to be capable of having a defined representation. To
achieve this, rkyv has a special "strict" mode that forces all of its
"archive" types to have `repr(C)`. This makes it so we'll be able to
(safely) deserialize data even if it was created by a version of puffin
that was built with a different version of the Rust compiler as the one
that created it. (There are alternatives to this approach, e.g., we
could invalidate puffin's entire cache based on whether it was built by
the same version of the Rust compiler that generated the cached data,
but this seems unfortunate. We'd like to start conservative I think.)

The `repr(C)` is needed because, by default, Rust types do not have a
defined ABI. The main downside of `repr(C)` is that it will tend to use
more space than the native Rust representation. For how we want to use
`rkyv`, this will translate to bigger cache sizes and more memory usage.
But I don't expect the difference to be enormous.

This is usually enough, but not for some types such as Rust tuples that
neither have any defined representation and don't have a way of making
it defined. Therefore, rkyv's `strict` mode is incompatible with the use
of tuples. And thus, this change from a tuple to a named type.

(I had been previously tempted to do this anyway for stylistic reasons
in my previous foray with this crate, but didn't have a compelling
enough reason to do so.)