pypi: ensure "files" retrieved from PyPI are given a stable order
Previously, we *may* have been depending on PyPI to send us a list of
files for a package in a stable order. While some tests do not need to
suggest that we were actually depending on the order (or even that PyPI
has an unstable order), we sort the results anyway to provide a
guarantee.