types: change BaseUrl in FileLocation to a String
The motivation for this is that it will make it easier to derive rkyv's
traits. Ideally, the `url` crate would add optional rkyv support (like
it has serde support), but rkyv hasn't really reached the maturity level
for that yet I think.

We could also try to convince the rkyv project maintainers to support
this[1] like they do for other crates, but I'm not sure it's possible
since the `url` crate (understandably) doesn't really expose its
internal representation. (Although perhaps its `deserialize_internal`
routines could be used to hack around this, I'm not sure that it works
for rkyv. Getting an archive type requires knowledge about the
representation in the type system. Needing to go through a function call
for it won't work.)

We tell ourselves this is OK because of the following:

1. Creating a `FileLocation::RelativeUrl` is believe to be rare in the
   first place.
2. In all cases where `BaseUrl` was used, it was to join it with another
   component. That requires parsing the other component. So this change
   is going from 1 URL parse call to 2 URL parse calls when
   `BaseUrl::join_relative` is called. Not the best, but perhaps
   different than 0 calls to 1 call. And as in (1), this is rare anyway,
   so the cost should be minimal.

[1]: https://github.com/rkyv/rkyv/blob/6892a9cd5d6a78c8d753c3dd35bfd814b00afa69/rkyv/Cargo.toml#L26-L37