distribution-types: derive rkyv traits for FileLocation
Since PathBuf (and OsString) are platform dependent, it's a little
tricky to deal with them in a zero-copy context. rkyv provides
AsString[1] for dealing with them as if they were strings, which is what
we use here. But this comes with the downside that serialization will
fail if the path is not valid UTF-8. If that winds up being an issue in
this specific case, we can probably switch to a `Vec<u8>`. I think it's
worth trying this approach first though.

[1]: https://docs.rs/rkyv/latest/rkyv/with/struct.AsString.html