Improve retry of pip install scenario tests (#1063)
Follow-up to #996 with retries of the full `pip_install_scenario` module
since there are a few tests that flake and tracking down the
non-determinism is proving to be quite complicated.

Adds a profile that can be used to disable retries, e.g. `--profile
no-retry` when you expect the scenario to change.

As before, hopefully this is a short-term measure.