distribution-types: add rkyv impls for File
We achieve this by replacing chrono's DateTime<Utc> type with a
millisecond UTC timestamp. Surprisingly, Chrono actually has optional
support for rkyv, but it's a little awkward to use. It doesn't, for
example, derive Debug impls. Also, the in-progress Chrono work (0.5.x)
doesn't seem to carry over rkyv's "validation" feature.

Anyway, this just isn't worth fussing about because we only use this
field for simple comparisons when building a VersionMap.